+++ packages/sources/react/src/components/forms/VtmnSelect/VtmnSelect.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
-import clsx from 'clsx';
 import '@vtmn/css-select/dist/index-with-vars.css';
+import clsx from 'clsx';
+import React from 'react';
 
 export interface VtmnSelectProps extends React.ComponentPropsWithRef<'select'> {
   error?: boolean;
@@ -18,7 +18,6 @@ export const VtmnSelect = ({
   id,
   labelText,
   options = [],
-  ref,
   ...props
 }: VtmnSelectProps) => {
   const errorTextId = `error-text-${id}`;