+++ packages/sources/svelte/src/guidelines/iconography/VtmnIcon/test/VtmnIcon.spec.js
@@ -1,6 +1,5 @@
 import '@testing-library/jest-dom';
-
-import { render, fireEvent } from '@testing-library/svelte';
+import { render } from '@testing-library/svelte';
 import VtmnIcon from '../VtmnIcon.svelte';
 
 describe('VtmnIcon', () => {