+++ packages/sources/vue/src/components/index.ts
@@ -7,6 +7,9 @@ import VtmnLoader from './indicators/VtmnLoader';
 import VtmnPrice from './indicators/VtmnPrice';
 import VtmnTag from './indicators/VtmnTag';
 import VtmnRating from './indicators/VtmnRating';
+import VtmnBreadcrumb from './navigation/VtmnBreadcrumb';
+import VtmnBreadcrumbItem from './navigation/VtmnBreadcrumbItem';
+import VtmnNavbar from './navigation/VtmnNavbar';
 import VtmnSearch from './navigation/VtmnSearch';
 import VtmnAlert from './overlays/VtmnAlert';
 import VtmnModal from './overlays/VtmnModal';
@@ -33,6 +36,9 @@ export {
   VtmnPrice,
   VtmnTag,
   VtmnRating,
+  VtmnBreadcrumb,
+  VtmnBreadcrumbItem,
+  VtmnNavbar,
   VtmnSearch,
   VtmnAlert,
   VtmnModal,