+++ packages/sources/css/presets/tailwind/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.33.110](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.109...@vtmn/css-tailwind-preset@0.33.110) (2022-05-09)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
+## [0.33.109](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.108...@vtmn/css-tailwind-preset@0.33.109) (2022-05-09)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
+## [0.33.108](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.107...@vtmn/css-tailwind-preset@0.33.108) (2022-05-09)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
 ## [0.33.107](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.106...@vtmn/css-tailwind-preset@0.33.107) (2022-05-09)
 
 **Note:** Version bump only for package @vtmn/css-tailwind-preset