+++ packages/sources/css/src/components/indicators/loader/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.25](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-loader@0.2.24...@vtmn/css-loader@0.2.25) (2022-05-09)
+
+**Note:** Version bump only for package @vtmn/css-loader
+
+
+
+
+
 ## [0.2.24](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-loader@0.2.23...@vtmn/css-loader@0.2.24) (2022-04-21)
 
 **Note:** Version bump only for package @vtmn/css-loader