+++ packages/sources/css/src/components/indicators/price/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-price@0.3.1...@vtmn/css-price@0.3.2) (2022-05-09)
+
+**Note:** Version bump only for package @vtmn/css-price
+
+
+
+
+
 ## [0.3.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-price@0.3.0...@vtmn/css-price@0.3.1) (2022-04-21)
 
 **Note:** Version bump only for package @vtmn/css-price