+++ packages/sources/css/src/components/overlays/tooltip/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.15](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tooltip@0.6.14...@vtmn/css-tooltip@0.6.15) (2022-05-09)
+
+**Note:** Version bump only for package @vtmn/css-tooltip
+
+
+
+
+
 ## [0.6.14](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tooltip@0.6.13...@vtmn/css-tooltip@0.6.14) (2022-04-21)
 
 **Note:** Version bump only for package @vtmn/css-tooltip