+++ packages/sources/css/src/components/structure/list/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.19](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-list@0.2.18...@vtmn/css-list@0.2.19) (2022-05-09)
+
+**Note:** Version bump only for package @vtmn/css-list
+
+
+
+
+
 ## [0.2.18](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-list@0.2.17...@vtmn/css-list@0.2.18) (2022-04-21)
 
 **Note:** Version bump only for package @vtmn/css-list